refactor: rename ListTemplate to be consistent with new named files
