+++ pages/_main/index.vue
@@ -19,23 +19,23 @@
       </template>
     </DHero>
     <!-- The help page is a list of simple cards, informed by .yml files -->
-    <ListTemplate v-if="$route.params.main === 'help'" :data="data" />
+    <FilesToCards v-if="$route.params.main === 'help'" :data="data" />
     <!-- The about page is a made of sections, informed by .md files -->
     <About v-else-if="$route.params.main === 'about'" :data="data" />
     <!-- This is for pages with subdirectories. They are a list of categories -->
-    <ListTemplate v-else :data="list" />
+    <FilesToCards v-else :data="list" />
   </div>
 </template>
 
 <script>
 import { DHero } from '@brown-ccv/disco-vue-components';
-import ListTemplate from '@/components/blocks/ListTemplate.vue';
+import FilesToCards from '@/components/blocks/FilesToCards.vue';
 import About from '@/components/blocks/About.vue';
 
 export default {
   components: {
     DHero,
-    ListTemplate,
+    FilesToCards,
     About
   },
   async asyncData({ $content, params }) {
